chore: sync conversation model between the two projects
